Move critical classes out of init
Move Receptor and Connection classes into the proper place
